% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokenstick-package.R
\docType{package}
\name{brokenstick-pkg}
\alias{brokenstick-pkg}
\title{\pkg{brokenstick}: A package for irregular longitudinal data.}
\description{
The broken stick model describes a set of individual curves
by a linear mixed model using first order linear B-splines. The
main use of the model is to align irregularly observed data to a
user-specified grid of break ages.
}
\details{
The \pkg{brokenstick} package contains functions for
fitting a broken stick model to data, for predicting broken
stick curves for new data, and for plotting the results.
}
\note{
Development of this package was kindly supported under the Healthy
Birth, Growth and Development knowledge integration (HBGDki)
program of the Bill & Melinda Gates Foundation.
}
\section{brokenstick functions}{

The main functions are:
\tabular{ll}{
\code{brokenstick()} \tab Fit a broken stick model to irregular data\cr
\code{predict()} \tab Obtain predictions from fitted model\cr
}
}

\references{
van Buuren, S. (2018). \emph{Flexible Imputation of Missing Data. Second Edition}. Chapman & Hall/CRC. Chapter 11.
\url{https://stefvanbuuren.name/fimd/sec-rastering.html#sec:brokenstick}
}
\seealso{
\code{\link{brokenstick}},
\code{\link{EB}}, \code{\link{predict.brokenstick}}
}
