% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.brokenstick}
\alias{coef.brokenstick}
\title{Extract Model Coefficients from brokenstick Object}
\usage{
\method{coef}{brokenstick}(
  object,
  complete = TRUE,
  ...,
  hide = c("right", "left", "boundary", "internal", "none")
)
}
\arguments{
\item{object}{A \code{brokenstick} object}

\item{complete}{for the default (used for \code{lm}, etc) and
    \code{aov} methods:
    logical indicating if the full coefficient vector should be returned
    also in case of an over-determined system where some coefficients
    will be set to \code{\link{NA}}, see also \code{\link[stats]{alias}}.  Note
    that the default \emph{differs} for \code{\link[stats]{lm}()} and
    \code{\link[stats]{aov}()} results.}

\item{...}{other arguments.}

\item{hide}{Should output for boundary knots be hidden in the print,
summary and plot functions? Can be \code{"right"}, \code{"left"}, \code{"boundary"},
\code{"internal"} or \code{"none"}.
If not specified, it is read from \code{object$hide}.}
}
\description{
Extract Model Coefficients from brokenstick Object
}
