% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditional.means.R
\name{conditional.means}
\alias{conditional.means}
\title{Calculated broken stick estimates from a fitted model}
\usage{
conditional.means(fit)
}
\arguments{
\item{fit}{The fitted model of class \code{brokenstick},
as produced by \code{brokenstick()}}
}
\value{
Matrix with conditional means. Each row corresponds to
a subject, each column corresponds to an (internal) knot.
}
\description{
The broken stick estimates are the conditional means of the
posterior distribution of the random effect. The means are
calculated as the sum of the fixed and random
effects. This function returns this sum at every subject
at every breakpoint.
}

