% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_brokenstick.R
\name{control_brokenstick}
\alias{control_brokenstick}
\title{Set controls to steer calculations}
\usage{
control_brokenstick(
  lmer = lmerControl(check.nobs.vs.nRE = "warning"),
  kr = control_kr(),
  na.action = na.exclude
)
}
\arguments{
\item{lmer}{A list generated by \link[lme4:lmerControl]{lme4::lmerControl}. The default
is set to \code{lmerControl(check.nobs.vs.nRE = "warning")}, which turn
fatal errors with respect the number of parameters into warnings. Use
\code{lmerControl(check.nobs.vs.nRE = "ignore")} to silence \code{lmer()}.}

\item{kr}{A list generated by \link{control_kr}.}

\item{na.action}{The function to call for the \code{na.action} argument.
The default is \code{na.exclude}.}
}
\description{
Set controls to steer calculations
}
