% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_brokenstick.R
\name{control_kr}
\alias{control_kr}
\title{Set controls for Kasim-Raudenbush sampler}
\usage{
control_kr(
  model = c("argyle", "cole", "none"),
  runin = 100L,
  ndraw = 200L,
  par_skip = 1L,
  imp_skip = Inf
)
}
\arguments{
\item{model}{Correlation model: \code{argyle} (default), \code{cole} or \code{none}}

\item{runin}{Number of run-in iterations}

\item{ndraw}{Number of parameter draws}

\item{par_skip}{Number of iterations to next parameter draw}

\item{imp_skip}{Number of iterations to next outcome draw}
}
\value{
A list with five components
}
\description{
Set controls for Kasim-Raudenbush sampler
}
