% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smocc.hgtwgt.R
\docType{data}
\name{fit.hgt}
\alias{fit.hgt}
\title{Broken stick model, fitted object}
\format{An object of class \code{lmerMod}, fitted by the
\code{lmer()} function in the \pkg{lme4} package. If fitted by
\code{fit.brokenstick()}, the object contains three additional
attributes. These are:
\describe{
	\item{knots}{A vector of internal knots used by \code{make.basis()}}
	\item{Boundary.knots}{The boundary knots of the linear spline}
	\item{model}{A string with value \code{"brokenstick"}}
	}}
\description{
The object \code{fit.hgt} is an object of class \code{lmerMod}, with three additional attributes.
}
\note{
This dataset is property of the Netherlands Organisation for Applied Scientific Research TNO. Distribution is not permitted.
Inquiries at \email{stef.vanbuuren@tno.nl}.
}
\seealso{
\code{\link[lme4]{lmer}}, \code{\link[lme4]{merMod-class}},
\code{\link{fit.brokenstick}}
}
\keyword{datasets}

