% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smocc_hgtwgt.R
\docType{data}
\name{fit_1933}
\alias{fit_1933}
\title{Broken stick model fitted on 1933 children}
\format{An object of class \code{brokenstick}, fitted by the
\code{brokenstick()}. This is an \code{lmerMod} object with additional
slots:
\describe{
\item{knots}{A vector of internal knots used by \code{make_basis()}}
\item{boundary}{The boundary knots of the linear spline}
\item{degree}{The degree of the spline. The brokenstick model corresponds to \code{degree = 1}.}
\item{bs.call}{An object of class \code{call} representing the call that was used to construct the object.}
\item{xy}{A data frame with column names \code{subjid}, \code{x} and \code{y} containing the data on which the model was fitted when called by the call in slot \code{bs.call}.}
}}
\description{
The object \code{fit_1933} is an object of class \code{brokenstick} that contains the fitted broken stick model. The class \code{brokenstick} extend the \code{lmerMod} class with additional attributes.
}
\examples{
\dontrun{
# code used to construct the fit_1933 object
data <- smocc_hgtwgt
# remove outliers and records with missing ages and/or heights
trim <- with(data, !is.na(agedays) & !is.na(haz) & haz > (-5) & haz < 5)
d <- data[trim, ]
# fit the brokenstick model (takes substantial time)
knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24)/12, 4)
boundary <- c(0, 3)
d$subjid <- as.factor(d$subjid)
fit_1933 <- with(d, brokenstick(y = haz, x = age,
                                subjid = subjid,
                                knots = knots,
                                boundary = boundary))
}
}
\seealso{
\code{\link[lme4]{lmer}}, \code{\link[lme4]{merMod-class}},
\code{\link{brokenstick}}
}
\keyword{datasets}

