% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smocc_200.R
\docType{data}
\name{fit_200_light}
\alias{fit_200_light}
\title{Broken stick model with nine lines for 200 children (light)}
\format{
An object of class \link[=brokenstick-class]{brokenstick}, fitted by the
\code{\link[=brokenstick]{brokenstick()}}.
}
\description{
Object \code{fit_200_light} has class \code{brokenstick} and stores the
the model settings and parameter estimates.
}
\details{
The datasets was constructed as

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24)/12, 4)
fit_200_light <- brokenstick(hgt_z ~ age | id, data = smocc_200,
                       knots = knots, boundary = c(0, 3),
                       light = TRUE, seed = 1)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
