% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brokenstick.R
\name{fit_brokenstick}
\alias{fit_brokenstick}
\title{Fit a broken stick model to irregular data}
\usage{
fit_brokenstick(
  data,
  formula,
  knots = NULL,
  Boundary.knots = NULL,
  k = NULL,
  subset,
  weights,
  na.action = na.exclude,
  method = c("lmer", "kr", "model.frame"),
  control = list(),
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} with the data}

\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{knots}{optional, numerical vector with the locations of the breaks to be
placed on the values of \code{x}. Be careful with values outside the range
of the data since this extends the \code{boundary} knots (see below) beyond
the data range.}

\item{Boundary.knots}{optional, but recommended. Numerical vector of length 2 with the minimum and maximum
knot. This \code{boundary} setting is passed to \code{splines::bs()} as the
\code{Boundary.knots} argument. If not specified, then the range of \code{x}
is taken. If \code{knots} is specified, then the boundary range is extended
to include at least \code{range{knots}}.}

\item{k}{optional, scalar indicating the number of internal knots. If specified, then
\code{k} internal knots are placed at equidense quantiles of \code{x}. For example,
specifying \code{k = 1} puts a knot at the 50th quantile (median), specifying \code{k = 3} puts knots
at the 25th, 50th and 75th quantiles of \code{x}, and so on. If both \code{k} and
\code{knots} are specified, then \code{k} take precendence. Note that knots specified
via \code{k} are data-dependent and do not transfer well to other data sets. Use \code{knots}
to specify knots that are independent of the data \code{x}.}

\item{subset}{an optional vector specifying a subset of observations
    to be used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
    process.  Should be \code{NULL} or a numeric vector.
    If non-NULL, weighted least squares is used with weights
    \code{weights} (that is, minimizing \code{sum(w*e^2)}); otherwise
    ordinary least squares is used.  See also \sQuote{Details},}

\item{na.action}{The function to call for the \code{na.action} argument
in \code{lmer()}. The default is \code{na.exclude}.}

\item{method}{Either \code{"kr"} or \code{"lmer"}}

\item{control}{A list with argument that can be used to control the
fitting of \code{lmer()}. The default
is set to \code{lmerControl(check.nobs.vs.nRE = "warning")}, which turn
fatal errors with respect the number of parameters into warnings.}

\item{\dots}{Additional arguments passed down to \code{lmer()}
(e.g. to specify additional \code{lmer()} options.}
}
\value{
A fitted model of class \code{brokenstick}, which extends the
 class \code{lmerMod}
}
\description{
The broken stick model models an irregularly observed series
of measurement by scaling them onto a user-specified set of
'ideal' ages. The model codes age by a series of linear B-splines.
Differences between persons are expressed by a random effect
pertaining to each knot. On the individual level, each
modeled growth curve connect straight lines that join at the
chosen break ages, and hence look like a 'broken stick'.
}
\details{
Relations over time are modeled by the variance-covariance
parameters of the random effects. Currently, this matrix is estimated
as unstructured by \code{lmer()} from the \code{lme4} package.
This estimate may be unstable if
the number of children is small relative to the number of specified
knots.

This function can be time consuming for data sets with thousands of
children.
}
\examples{
data <- brokenstick::smocc_200
# fit <- fit_brokenstick(data, hgt.z ~ age | subjid, 0:2)
mf <- fit_brokenstick(smocc_200, hgt.z ~ age | subjid, knots = 0:2)

}
