% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smocc_hgtwgt.R
\docType{data}
\name{fit_hgt}
\alias{fit_hgt}
\title{Broken stick model, fitted object}
\format{An object of class \code{brokenstick}, fitted by the 
\code{brokenstick()}. This is an \code{lmerMod} object with two additional
slots: 
\describe{
\item{knots}{A vector of internal knots used by \code{make_basis()}}
\item{boundary}{The boundary knots of the linear spline}
}}
\description{
The object \code{fit_hgt} is an object of class \code{lmerMod}, with three additional attributes.
}
\note{
This dataset is property of the Netherlands Organisation for Applied Scientific Research TNO. Distribution is not permitted. 
Inquiries at \email{stef.vanbuuren@tno.nl}.
}
\seealso{
\code{\link[lme4]{lmer}}, \code{\link[lme4]{merMod-class}},
\code{\link{brokenstick}}
}
\keyword{datasets}

