% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.brokenstick}
\alias{fitted.brokenstick}
\title{Calculate fitted values}
\usage{
\method{fitted}{brokenstick}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{newdata}{Optional. A data frame in which to look for variables with
which to predict. The training data are used if omitted and
if \code{object$light} is \code{FALSE}.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
A numerical vector with predictions. The number of elements equals the
number of rows in \code{newdata}. If \code{newdata} is not specified, the function
looks for the training data in \code{object} as the element named \code{data}.
}
\description{
Calculate fitted values
}
\seealso{
Other brokenstick: 
\code{\link{residuals.brokenstick}()}
}
\concept{brokenstick}
