% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.brokenstick.values.R
\name{get.brokenstick.values}
\alias{get.brokenstick.values}
\title{Calculated broken stick estimates from a fitted model}
\usage{
get.brokenstick.values(fit)
}
\arguments{
\item{fit}{The fitted model of class \code{lmerMod}, presumably fitted
by \code{fit.brokenstick()}}
}
\value{
Matrix with broken stick values
}
\description{
The broken stick estimates are simply the sum of the fixed and random
effect. This function return that sum for all levels.
}

