% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get.knots}
\alias{get.knots}
\title{Obtain the knots from a broken stick model}
\usage{
get.knots(object, include.boundaries = FALSE)
}
\arguments{
\item{object}{An object of class \code{brokenstick} or \code{brokenstick.export}}

\item{include.boundaries}{A logical specifying whether the right-most (boundary) knots should be included. The default is \code{FALSE}, which return only the internal knots.}
}
\value{
A vector with knot locations
}
\description{
Obtain the knots from a broken stick model
}
\examples{
get.knots(fit.hgt)
}

