% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data}
\alias{get_data}
\title{Obtain the data from a broken stick model}
\usage{
get_data(object, ids = NULL)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}

\item{ids}{A vector specifying the id's of the persons. If omitted,
all id's are included.}
}
\value{
A data frame with x, y and z. The result is \code{NULL}
if \code{object} is not of class \code{brokenstick}.
}
\description{
Obtain the data from a broken stick model
}
\examples{
get_data(fit_200, ids = c(10001, 10002))
}
