% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_knots}
\alias{get_knots}
\title{Obtain the knots from a broken stick model}
\usage{
get_knots(
  object,
  whatknots = c("all", "internal", "boundary", "dropfirst", "droplast"),
  what = "all"
)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}

\item{whatknots}{A character vector of length 1 specifies the knot set.
Valid values are \code{"all"}, \code{"internal"}, \code{"boundary"},
\code{"dropfirst"} and \code{"droplast"}. The default is
\code{whatknots = "all"}}

\item{what}{Deprecated. Use \code{whatknots} instead.}
}
\value{
A vector with knot locations, either both, internal only or
boundary only. The result is \code{NULL} if \code{object} does not
have proper class. Returns \code{numeric(0)} if
there are no internal knots.
}
\description{
Obtain the knots from a broken stick model
}
\examples{
get_knots(fit_200, "internal")
}
