% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_label}
\alias{get_label}
\title{Get the label of a HBGD standard variable}
\usage{
get_label(x)
}
\arguments{
\item{x}{An character object with the name of one variable}
}
\value{
A character vector, either the label found in the table, or the name of the variable.
}
\description{
This functions consults the \code{hbgd::hbgd_labels} table to find the variable
label.
}
\note{
This function should be placed in the \code{hbgd} package.
}

