% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pev}
\alias{get_pev}
\title{Obtain proportion of explained variance from a broken stick model}
\usage{
get_pev(object)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}
}
\value{
Proportion of explained variance
}
\description{
Obtain proportion of explained variance from a broken stick model
}
\examples{
# get_pev(fit_200)
}
