% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_r2}
\alias{get_r2}
\title{Obtain proportion of explained variance from a broken stick model}
\usage{
get_r2(object, new_data)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}

\item{new_data}{Data on which \code{r.squared} must be calculated}
}
\value{
Proportion of explained variance
}
\description{
Obtain proportion of explained variance from a broken stick model
}
\examples{
get_r2(fit_200, smocc_200)
}
