% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_X}
\alias{get_X}
\title{Obtain the X model matrix from a broken stick model}
\usage{
get_X(object, ids = NULL)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}

\item{ids}{A vector specifying the id's of the persons. If omitted, all id's are included.}
}
\value{
A matrix with number of columns equal to the number of knots.
The result is \code{NULL} if \code{object} is not of class \code{brokenstick}.
}
\description{
Obtain the X model matrix from a broken stick model
}
\examples{
get_X(fit_1933, ids = c(10001, 10002))
}
