% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_y}
\alias{get_y}
\title{Obtain the y variable from a broken stick model}
\usage{
get_y(object, ids = NULL)
}
\arguments{
\item{object}{An object of class \code{brokenstick}}

\item{ids}{A vector specifying the id's of the persons. If omitted, all id's are included.}
}
\value{
A vector
The result is \code{NULL} if \code{object} is not of class \code{brokenstick}.
}
\description{
Obtain the y variable from a broken stick model
}
\examples{
get_y(fit_hgt, ids = c(10001, 10002))
}

