% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula for brokenstick model}
\usage{
parse_formula(f)
}
\arguments{
\item{f}{formula object}
}
\value{
A \code{list} with elements \code{x}, \code{y} and \code{g}.
Each element has length 1.
}
\description{
A bare bones formula parser to extract variables names from
formulas of \code{y ~ x | g}. It return the name of
the first variable mentioned in each formula component.
}
\examples{
# examples that yield identical result
brokenstick:::parse_formula(y ~ x | z)
brokenstick:::parse_formula(y + a ~ x + x1 | z + b)
brokenstick:::parse_formula(y + a + log(b) ~ x + x1 * c | z + d)
}
\author{
Stef van Buuren, 2020
}
