% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brokenstick}
\alias{plot.brokenstick}
\title{Plot observed and fitted trajectories by group}
\usage{
\method{plot}{brokenstick}(
  x,
  new_data,
  ...,
  .x = NULL,
  group = NULL,
  n_plot = 3,
  x_trim = c(-Inf, Inf),
  show = c(TRUE, TRUE)
)
}
\arguments{
\item{x}{An object of class \code{brokenstick}.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{...}{Extra arguments passed down to \code{\link[=predict_new.brokenstick]{predict_new.brokenstick()}}
and \code{\link[=plot_trajectory]{plot_trajectory()}}.}

\item{.x}{The \code{x} argument of the \code{\link[=predict_new.brokenstick]{predict_new.brokenstick()}} function.}

\item{group}{A vector with group identifications}

\item{n_plot}{A integer indicating the number of individual plots.
The default is 3, which plots the trajectories of the first three
groups. The \code{n_plot} is a safety measure to prevent unintended
plots of the entire data set.}

\item{x_trim}{A range on the x-axis that can be used to subset values
that are plotted.}

\item{show}{A logical vector of length 2. The first element specifies
whether the observed data are plotted, the second element specifies
whether the added data (e.g. knots) are plotted. The default is
\code{c(TRUE, TRUE)}.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
The \code{plot} method for a \code{brokenstick} object plots the observed and
fitted trajectories of one or more groups.
}
\examples{
library("brokenstick")
dat <- smocc_200
# fit one line model for data exploration
fit <- brokenstick(hgt ~ age | id, dat)

# plot measurements for first three cases
# plot(fit_200, smocc_200, zband = FALSE, show = c(TRUE, FALSE))

# fit model with knots at 1 and 2 years
#fit <- fit_brokenstick(dat, hgt.z ~ age | id, knots = 1:2)

#plot(fit, xlim = c(0, 2.2), ylim = c(-3, 3))
#plot(fit, ids = c(10005, 10012), xlim = c(0, 2.2))

# bokeh plots with xlim and ylim
#plot(fit, xlim = c(0, 2), ylim = c(-3, 3))
#plot(fit, ids = 10005)
}
\seealso{
\link{predict_new.brokenstick}, \link{plot_trajectory}.
}
\author{
Stef van Buuren 2020
}
