% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.brokenstick.export}
\alias{predict.brokenstick.export}
\title{Predict growth curve according to the broken stick model}
\usage{
\method{predict}{brokenstick.export}(object, y, age, type = "atknots",
  output = "long", include.boundaries = FALSE, onlynew = TRUE, ...)
}
\arguments{
\item{object}{An object of 
containing the estimated parameters of the broken stick model}

\item{y}{A vector with measurements}

\item{age}{A vector with decimal ages of length \code{length(y)}}

\item{type}{If \code{type = "curve"} (the default) 
the function returns the broken stick estimates. 
If \code{type = "response"}, the function returns a predicted value 
for each element of \code{y}.}

\item{\dots}{Additional arguments (not used)}
}
\value{
A numeric vector of length \code{length(slot(object, "knots"))
+ slot(object, "degree")} with 
predicted values
}
\description{
For a given child, extract the conditional modes of the
random effects as specified by the broken stick (=linear
piecewise spline) model.
}
\author{
Stef van Buuren
}

