% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.brokenstick}
\alias{predict.brokenstick}
\title{Predict growth curve according to the broken stick model}
\usage{
\method{predict}{brokenstick}(model, y, age)
}
\arguments{
\item{model}{An object of either class \code{brokenstick} or
of \code{brokenstick.export}
containing the estimated parameters of the broken stick model}

\item{y}{A vector with measurements}

\item{age}{A vector with decimal ages of length \code{length(y)}}
}
\value{
A data frame with \code{length(slot(model, "knots"))
+ slot(model, "degree")} elements with
predicted values
}
\description{
For a given child, extract the conditional modes of the
random effects as specified by the broken stick (=linear
piecewise spline) model.
}
\author{
Stef van Buuren, 2015
}

