% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictbs.R
\name{predictbs}
\alias{predictbs}
\title{Predict growth curve according to the broken stick model}
\usage{
predictbs(model, yname, y, age, sex, ga = 40, zscale = FALSE)
}
\arguments{
\item{model}{An object of either class \code{lmerMod} or
class \code{brokenstick.export}
containing the estimated parameters of the broken stick model
for the type of measurement in \code{yname}}

\item{yname}{A string (\code{"hgt"}, \code{"wgt"} or \code{"hdc"})
identifying the type of measurement}

\item{y}{A vector with measurements belonging to one person}

\item{age}{A vector with decimal ages of length \code{length(y)}}

\item{sex}{Either \code{"male"} or \code{"female"}}

\item{ga}{Numeric, gestational age in weeks.
The default is 40 weeks, term birth.}

\item{zscale}{A logical indicating whether the result should be in
the Z-scale (\code{zscale = TRUE}) or Y-scale (\code{zscale = FALSE}).
The default is \code{FALSE}.}
}
\value{
A \code{vector} with a length equal to the number of break
points (i.e. number of unique knots)
}
\description{
For a given child, extract the conditional modes of the
random effects as specified by the broken stick (=linear
piecewise spline) model. Modeling is done in the scale of the
Z-score. The code assumes that the estimated parameters of the
relevant broken stick model (fixed effect, variance-covariance of
random effect, and residual variance) are available in a special object
of class \code{brokenstick.export}.
}
\author{
Stef van Buuren, 2015
}

