% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.brokenstick}
\alias{print.brokenstick}
\title{Print brokenstick object}
\usage{
\method{print}{brokenstick}(
  x,
  digits = getOption("digits"),
  ...,
  hide = c("right", "left", "boundary", "internal", "none")
)
}
\arguments{
\item{x}{A \code{brokenstick} object}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link[base]{print.default}}.}

\item{...}{further arguments passed to or from other methods.}

\item{hide}{Should output for boundary knots be hidden in the print,
summary and plot functions? Can be  \code{"right"}, \code{"left"},\code{"boundary"},
\code{"internal"} or \code{"none"}.
If not specified, it is read from the field \code{x$hide}.}
}
\description{
Print brokenstick object
}
