% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.brokenstick}
\alias{residuals.brokenstick}
\title{Extract residuals from brokenstick model}
\usage{
\method{residuals}{brokenstick}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{newdata}{Optional. A data frame in which to look for variables with
which to predict. The training data are used if omitted and
if \code{object$light} is \code{FALSE}.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
A numerical vector with residuals The number of elements equals the
number of rows in \code{newdata}. If \code{newdata} is not specified, the function
looks for the training data in \code{object} as the element named \code{data}.
}
\description{
Extract residuals from brokenstick model
}
\seealso{
Other brokenstick: 
\code{\link{fitted.brokenstick}()}
}
\concept{brokenstick}
