% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_control.R
\name{set_control}
\alias{set_control}
\title{Set controls to steer calculations}
\usage{
set_control(
  method = c("kr", "lmer"),
  kr = control_kr(),
  lmer = lmerControl(check.nobs.vs.nRE = "warning")
)
}
\arguments{
\item{method}{String indicating estimation method: \code{"kr"} or \code{"lmer"}}

\item{kr}{A list generated by \link{control_kr}.}

\item{lmer}{A list generated by \link[lme4:lmerControl]{lme4::lmerControl}. The default
is set to \code{lmerControl(check.nobs.vs.nRE = "warning")}, which turns
fatal errors with respect the number of parameters into warnings. Use
\code{lmerControl(check.nobs.vs.nRE = "ignore")} to silence \code{lmer()}.}
}
\description{
Set controls to steer calculations
}
