% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smocc_200.R
\docType{data}
\name{smocc_200}
\alias{smocc_200}
\title{Infant growth of 0-2 years, SMOCC data extract}
\format{
A tibble with 1940 rows and 7 columns:
\describe{
\item{subjid}{ID, unique \code{id} of each child (numeric)}
\item{age}{Decimal age, 0-2.12 years (numeric)}
\item{sex}{Sex, \code{"male"} or \code{"female"} (character)}
\item{ga}{Gestational age, completed weeks (numeric)}
\item{bw}{Birth weight in grammes (numeric)}
\item{hgt}{Height measurement in cm (34-102) (numeric)}
\item{hgt.z}{Height in SDS relative to WHO standard) (numeric)}
}
}
\source{
Herngreen WP, van Buuren S, van Wieringen JC, Reerink JD, Verloove-Vanhorick SP & Ruys JH (1994). Growth in length and weight from birth to 2 years of a representative sample of Netherlands children (born in 1988-89) related to socio-economic status and other background characteristics. \emph{Annals of Human Biology}, \bold{21}, 449-463.
}
\description{
Longitudinal height and weight measurements during ages 0-2 years
for a representative sample of 1933 Dutch children born in 1988-1989.
The dataset \code{smocc_200} is a subset of the full data covering
200 children.
}
\keyword{datasets}
