% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.brokenstick}
\alias{summary.brokenstick}
\title{Create summary of brokenstick object}
\usage{
\method{summary}{brokenstick}(
  object,
  ...,
  cor = FALSE,
  lower = TRUE,
  hide = c("right", "left", "boundary", "internal", "none")
)
}
\arguments{
\item{object}{A \code{brokenstick} object}

\item{...}{additional arguments affecting the summary produced.}

\item{cor}{Logical. Should the function return the correlation matrix
instead of the covariance matrix? The default is \code{FALSE}.}

\item{lower}{Logical. Print lower triangle of correlation/covariance
matrix?}

\item{hide}{Should output for boundary knots be hidden in the print,
summary and plot functions? Can be \code{"left"}, \code{"right"}, \code{"boundary"},
\code{"internal"} or \code{"none"}.
If not specified, it is read from the field \code{object$hide}.}
}
\description{
Create summary of brokenstick object
}
