% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcorr.R
\name{VarCorr}
\alias{VarCorr}
\title{Extract Variance and Correlation Components}
\usage{
VarCorr(x, scale = c("cov", "cor"), ...)
}
\arguments{
\item{x}{Object of class \code{brokenstick}}

\item{scale}{Either \code{"cov"} (default) for the covariance matrix,  or \code{"cor"}
for the correlation matrix.}

\item{\dots}{Not used.}
}
\description{
Extracts variance-covariance or correlation matrix from a
\code{brokenstick} object.
}
