% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokenstick-package.R
\docType{package}
\name{brokenstick-pkg}
\alias{brokenstick-pkg}
\title{\pkg{brokenstick}: A package for irregular longitudinal data.}
\description{
The broken stick model describes a set of individual curves
by a linear mixed model using first order linear B-splines. The
main use of the model is to align irregularly observed data to a
user-specified grid of break ages.
}
\details{
The \pkg{brokenstick} package contains functions for
fitting a broken stick
model to data, for exporting the parameters of the model for
independent use outside this package, and for predicting broken
stick curves for new data.
}
\note{
Development of this package was kindly supported under the Healthy
Birth, Growth and Development knowledge integration (HBGDki)
program of the Bill & Melinda Gates Foundation.
}
\section{brokenstick functions}{

The main functions are:
\tabular{ll}{
  \code{brokenstick()} \tab Fit a broken stick model to irregular data\cr
  \code{predict()} \tab Obtain predictions from fitted model\cr
  }
}

\references{
van Buuren, S. (2012). \emph{Flexible Imputation of Missing Data}. Chapman & Hall/CRC, 2012. Chapter 9.
}
\seealso{
\code{\link{brokenstick}},
\code{\link{EB}}, \code{\link{predict.brokenstick}}
}
