% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{brokenstick}
\alias{brokenstick}
\title{Fit a \code{brokenstick} model to irregular data}
\usage{
brokenstick(
  formula,
  data,
  knots = NULL,
  boundary = NULL,
  k = NULL,
  degree = 1L,
  method = c("lmer", "kr"),
  control = control_brokenstick(),
  seed = NA,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the outcome, the predictor and the group
variable in \code{data}. The generic shape is \code{formula = y ~ x | group}. The
left-hand side is the outcome, the right-hand side the predictor, and the
name of the grouping variable occurs after the \code{|} sign. Formula treatment
is non-standard: 1) \code{y} and \code{x} should be numeric, 2) only one variable
is allowed in each model term (additional variables will be ignored).}

\item{data}{A data frame or matrix containing the outcome (numeric),
predictor (numeric) and group (numeric, factor, character) variable.}

\item{knots}{Optional, but recommended. Numerical vector with the
locations of the internal knots to be placed on the values of the predictor.}

\item{boundary}{Optional, but recommended. Numerical vector of
length 2 with the left and right boundary knot. The \code{boundary}
setting is passed to \code{\link[splines:bs]{splines::bs()}} as the \code{Boundary.knots} argument.
If not specified, the function determines the boundary knots as
\code{range(x)}. When specified, the \code{boundary} range is internally
expanded to include at least \code{range(knots)}.}

\item{k}{Optional, a convenience parameter for the number of
internal knots. If specified, then \code{k} internal knots are placed
at equidense quantiles of the predictor. For example,
specifying \code{k = 1} puts a knot at the 50th quantile (median),
setting \code{k = 3} puts knots at the 25th, 50th and 75th quantiles,
and so on. If the user specifies both \code{k} and \code{knots} arguments
then \code{knots} takes precedence.}

\item{degree}{the degree of the spline. The broken stick model
requires linear splines, so the default is \code{degree = 1}.
Setting \code{degree = 0} yields (crisp) dummy coding, and one
column less than for \code{degree = 1}. The \code{brokenstick} package supports
only \code{degree = 0} and \code{degree = 1}.}

\item{method}{Estimation method. Either \code{"kr"} (for the
Kasim-Raudenbush sampler) or \code{"lmer"} (for \code{\link[lme4:lmer]{lme4::lmer()}}) (default).}

\item{control}{A list with parameters. Use \code{control_brokenstick()}
to generate.}

\item{seed}{Seed number for \code{\link[base:Random]{base::set.seed()}}. Use \code{NA} to bypass
seed setting.}

\item{\dots}{Not currently used, but required for extensibility.}
}
\value{
A \code{brokenstick} object.
}
\description{
The \code{brokenstick()} function fits an irregularly observed series
of measurements onto a user-specified grid of points (knots).
The model codes the grid by a series of linear B-splines.
Each modelled trajectory consists of straight lines that join at
the chosen knots and look like a broken stick. Differences between
observations are expressed by a random effect per knot.
}
\details{
The variance-covariance matrix of the random effects absorbs the
relations over time. By default, this matrix is estimated
as unstructured by \code{\link[lme4:lmer]{lme4::lmer()}}.
This estimate may be unstable if
the number of children is small relative to the number of specified
knots. The function can be time consuming for data sets with thousands of
children.

An alternative - often faster for models with many random effects -
is to use the Bayesian Kasim-Raudenbush sampler (method \code{kr}). That
method also allow for enforcing a simple structure on the
variance-covariance matrix of the random effects. Currently, there
are three such models: \code{argyle}, \code{cole} and \code{none}. See \code{\link[=kr]{kr()}} and
\code{\link[=control_brokenstick]{control_brokenstick()}} for more details.
}
\note{
Note that automatic knot specification is data-dependent, and may not reproduce
on other data. Likewise, knots specified via \code{k} are data-dependent and do not transfer
to other  data sets. Fixing the model requires specifying both \code{knots} and
\code{boundary}.
}
\examples{
data <- smocc_200[1:1198, ]
fit <- brokenstick(hgt.z ~ age | id, data, knots = 0:3)
plot(fit, data, n_plot = 9)

# using KR sampler
fit <- brokenstick(hgt.z ~ age | id,
  data = data, knots = 0:3,
  method = "kr", seed = 1
)
\donttest{
knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24) / 12, 4)
boundary <- c(0, 3)
fit_lmer <- brokenstick(hgt.z ~ age | id,
  data = data,
  knots = knots, boundary = boundary
)
fit_kr <- brokenstick(hgt.z ~ age | id,
  data = data, knots = knots,
  boundary = boundary, method = "kr"
)
}
}
