% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditional.means.R
\name{conditional.means}
\alias{conditional.means}
\title{Calculated broken stick estimates from a fitted model}
\usage{
conditional.means(fit, long = TRUE)
}
\arguments{
\item{fit}{The fitted model of class \code{brokenstick},
as produced by \code{brokenstick()}}

\item{long}{A logical indicating whether the result should be in the
long form (the default), or in the broad form.}
}
\value{
If \code{long} is \code{TRUE}, a matrix with three columns, representing the
 subject level, breakpoint value and the conditional mean of the
 distribution of broken stick values.
If \code{long} is \code{FALSE}, a matrix with
\code{length(knots) + degree} columns. Each row corresponds to
a subject, each column corresponds to a knot.
}
\description{
The broken stick estimates are the conditional means of the
posterior distribution of the random effect. The means are
calculated as the sum of the fixed and random
effects. This function returns this sum at every subject
at every breakpoint.
}

