% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_brokenstick.R
\name{control_brokenstick}
\alias{control_brokenstick}
\title{Set controls to steer calculations}
\usage{
control_brokenstick(
  method = c("lmer", "kr"),
  lmer = lmerControl(check.nobs.vs.nRE = "warning"),
  kr = list(runin = 100L, ndraw = 200L, par_skip = 1L, imp_skip = Inf),
  na.action = na.exclude
)
}
\arguments{
\item{method}{Either \code{"kr"} (for the Kasim-Raudenbush sampler)
or \code{"lmer"} (for \code{\link[lme4:lmer]{lme4::lmer()}}).}

\item{lmer}{A list generated by \link[lme4:lmerControl]{lme4::lmerControl}. The default
is set to \code{lmerControl(check.nobs.vs.nRE = "warning")}, which turn
fatal errors with respect the number of parameters into warnings. Use
\code{lmerControl(check.nobs.vs.nRE = "ignore")} to silence \code{lmer()}.}

\item{kr}{A list with the following components:
\itemize{
\item \code{runin}: Number of run-in iterations
\item \code{ndraw}: Number of parameter draws
\item \code{par_skip}: Number of iterations to next parameter draw
\item \code{imp_skip}: Number of iterations to next outcome draw
}}

\item{na.action}{The function to call for the \code{na.action} argument.
The default is \code{na.exclude}.}
}
\description{
Set controls to steer calculations
}
