% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_control.R
\name{control_kr}
\alias{control_kr}
\title{Set controls for Kasim-Raudenbush sampler}
\usage{
control_kr(
  cormodel = c("none", "argyle", "cole"),
  m = 0L,
  start = 101L,
  end = 300L,
  thin = 1L
)
}
\arguments{
\item{cormodel}{String indicating the correlation model:
\code{"none"} (default), \code{"argyle"} or \code{"cole"}}

\item{m}{Integer. Number of multiple imputations. The default \code{m = 0L}.}

\item{start}{Integer. The iteration number of the first observation}

\item{end}{Integer. The iteration number of the last observation}

\item{thin}{Integer. The thinning interval between consecutive observations}
}
\value{
\preformatted{    A list with five components
}
}
\description{
Set controls for Kasim-Raudenbush sampler
}
