% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_control.R
\name{control_kr}
\alias{control_kr}
\title{Set controls for Kasim-Raudenbush sampler}
\usage{
control_kr(
  n = 200L,
  m = 0L,
  start = 100L,
  thin = 1L,
  cormodel = c("none", "argyle", "cole")
)
}
\arguments{
\item{n}{Integer. Number of samples from posterior. Default:  \code{n = 200}.}

\item{m}{Integer. Number of multiple imputations. Default: \code{m = 0}.}

\item{start}{Integer. The iteration number of the first observation}

\item{thin}{Integer. The thinning interval between consecutive observations}

\item{cormodel}{String indicating the correlation model:
\code{"none"} (default), \code{"argyle"} or \code{"cole"}}
}
\value{
\preformatted{    A list with five components. The function calculates parameters
            `end` (the iteration number of the last iteration) and
            `thin_imp` (thinning factor for multiple imputations) from the
            other inputs.
}
}
\description{
Set controls for Kasim-Raudenbush sampler
}
