% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{EB}
\alias{EB}
\title{Empirical Bayes predictor for random effects}
\usage{
EB(model, y, X, Z = X, BS = FALSE)
}
\arguments{
\item{model}{An object of class \code{lmerMod} (typically created by
\code{fit.brokenstick()}) or class \code{brokenstick.export}
(typically generated by \code{export.brokenstick()}).}

\item{y}{A vector of new measurements for unit j, scaled in the same metric as the fitted model.}

\item{X}{A \code{nj * p} matrix with fixed effects for unit j, typically produced by \code{make.basis()}.}

\item{Z}{A \code{nj * q} matrix with random effects for unit j. The default sets \code{Z} equal to \code{X}.}

\item{BS}{A logical indicating whether broken stick estimates should be
returned (\code{BS = TRUE}) or the random effects (\code{BS = FALSE}).
The default is \code{FALSE}.}
}
\value{
A vector of length q containing the random effect or broken stick  estimates for unit j.
}
\description{
This function can estimate random effect for a given set of
model estimates and new user data. Contrary to many current
implementations for random effect prediction, the unit may be
new to the model. The methods implements the
EB estimate (also known as BLUP)
as described in Skrondral and Rabe-Hasketh, 2009, p. 683.
This function can also provide the broken stick estimate for a given level,
which is simply the sum of (global) fixed and (level) random effects
The current implementation does not provide prediction errors.
}
\examples{
library(mice)
data <- tbc[tbc$id < 1000 & tbc$age < 2.5,]
fit <- fit.brokenstick(z = data$hgt.z, age = data$age, id = data$id)
#
# conventional random effect for child id 8
ranef(fit)$id[1,]
#
# EB estimate random effect for child id 8
est <- export.brokenstick(fit)
y <- data[data$id == 8, "hgt.z"]
y
X <- make.basis(data[data$id == 8, "age"])
X
EB(est, y, X)
}
\author{
Stef van Buuren, 2015
}
\references{
Skrondal, A., Rabe-Hesketh, S. (2009).
Prediction in multilevel generalized linear models.
J. R. Statist. Soc. A, 172, 3, 659-687.
}

