% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{export.brokenstick}
\alias{export.brokenstick}
\title{Export the estimates of a fitted lmer() model}
\usage{
export.brokenstick(model)
}
\arguments{
\item{model}{An object of class \code{lmerMod} or class
\code{broken.stick.export} (typically generated
by a previous call to \code{export.brokenstick()}).}
}
\value{
A \code{list} of class \code{broken.stick.estimate}, with elements corresponding to the estimates parameters of the fitted model.
}
\description{
Exports the crucial estimates of a fitted lmer() model so that the
stored estimates can be used by the EB() function
to calculate random effect estimates for new user data.
}

