% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{fit.brokenstick}
\alias{fit.brokenstick}
\title{Fit a broken stick model to data}
\usage{
fit.brokenstick(yvar = "hgt.z", data)
}
\arguments{
\item{yvar}{a string (e.g. \code{"hgt"} or \code{"hgt.z"}) indicating the outcome variable}

\item{data}{a data frame with the data, including variables \code{x0} through
\code{x12}}
}
\value{
The fitted model of class \code{lmerMod}
}
\description{
This function fits the 0-3 years broken stick model to
the Z-score of the outcome. This function can be time consuming.
}
\examples{
library(mice)
data <- tbc[tbc$id < 200 & tbc$age < 2.5,]
fit <- fit.brokenstick("hgt.z", data)
plot(fit)
}

