% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.brokenstick}
\alias{fitted.brokenstick}
\title{Calculate fitted values}
\usage{
\method{fitted}{brokenstick}(object, new_data = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
See \code{\link[=predict_new.brokenstick]{predict_new.brokenstick()}}.
}
\description{
Calculate fitted values
}
\seealso{
Other brokenstick: 
\code{\link{residuals.brokenstick}()}
}
\concept{brokenstick}
