% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.brokenstick}
\alias{fitted.brokenstick}
\title{Calculate fitted values}
\usage{
\method{fitted}{brokenstick}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{newdata}{Optional. A data frame in which to look for variables with
which to predict. If omitted and if \code{isFALSE(object$light)}, the fitted
predictors are used.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
See \code{\link[=predict.brokenstick]{predict.brokenstick()}}.
}
\description{
Calculate fitted values
}
\seealso{
Other brokenstick: 
\code{\link{residuals.brokenstick}()}
}
\concept{brokenstick}
