% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{get.brokenstick.values}
\alias{get.brokenstick.values}
\title{Obtain broken stick values from a fitted model}
\usage{
get.brokenstick.values(fit)
}
\arguments{
\item{fit}{The fitted model of class \code{lmerMod}, presumably fitted
by \code{fit.brokenstick()}}
}
\value{
Matrix with broken stick values
}
\description{
Obtain broken stick values from a fitted model
}

