% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenstick.R
\name{get.brokenstick.values}
\alias{get.brokenstick.values}
\title{Calculated broken stick values from a fitted broken stick model}
\usage{
get.brokenstick.values(fit)
}
\arguments{
\item{fit}{The fitted model of class \code{lmerMod}, presumably fitted
by \code{fit.brokenstick()}}
}
\value{
Matrix with broken stick values
}
\description{
The broken stick values are simply the sum of the fixed and random
effect. This function return that sum for all levels.
}

