% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_knots}
\alias{get_knots}
\title{Obtain the knots from a broken stick model}
\usage{
get_knots(object)
}
\arguments{
\item{object}{An object of class \code{brokenstick} or \code{brokenstick.export}}
}
\value{
A vector with knot locations (both internal and boundary)
}
\description{
Obtain the knots from a broken stick model
}
\examples{
get_knots(fit_206)
}
