% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr.R
\name{kr}
\alias{kr}
\title{Kasim-Raudenbush sampler for two-level normal model}
\usage{
kr(y, x, g, control, na.action)
}
\arguments{
\item{y}{Vector with outcome value}

\item{x}{Matrix with predictor value}

\item{g}{Vector with ID values}

\item{control}{A list with elements \code{runin}, \code{ndraws} and \code{skip}}

\item{na.action}{Not really used here}
}
\value{
A list with results. see kr_vector
}
\description{
Simulates posterior distributions of parameters from a two-level
normal model with heterogeneous within-cluster variances
(Kasim and Raudenbush, 1998). Imputations can be drawn as an
extra step to the algorithm.
}
\references{
Kasim RM, Raudenbush SW. (1998). Application of Gibbs sampling to nested
variance components models with heterogeneous within-group variance. Journal
of Educational and Behavioral Statistics, 23(2), 93--116.

Van Buuren, S. (2011) Multiple imputation of multilevel data. In Hox, J.J.
and and Roberts, J.K. (Eds.), \emph{The Handbook of Advanced Multilevel
Analysis}, Chapter 10, pp. 173--196. Milton Park, UK: Routledge.
}
