% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr.R
\name{kr}
\alias{kr}
\title{Kasim-Raudenbush sampler for two-level normal model}
\usage{
kr(y, x, g, control, seed)
}
\arguments{
\item{y}{Vector with outcome value}

\item{x}{Matrix with predictor value}

\item{g}{Vector with group values}

\item{control}{A list created by \code{\link[=control_kr]{control_kr()}}}

\item{seed}{Seed number for \code{\link[base:Random]{base::set.seed()}}. Use \code{NA} to bypass
seed setting.}
}
\value{
A list with components:\preformatted{* `beta`  Fixed effects
* `omega` Variance-covariance of random effects
* `sigma2_j` Residual variance per group
* `sigma2` Average residual variance
* `draws` A matrix with `ndraw` columns with draws for missing data
}
}
\description{
Simulates posterior distributions of parameters from a two-level
normal model with heterogeneous within-cluster variances
(Kasim and Raudenbush, 1998). Imputations can be drawn as an
extra step to the algorithm.
}
\details{
The speed of the Kasim-Raudenbush sampler is almost
independent of the number of random effect, and foremost depends
on the \emph{total number of iterations}: \code{end} in the \code{control} list.

The defaults \code{start = 101}, \code{end = 300} and \code{thin = 1}
provide 200 draws with a \emph{reasonable} approximation to the variance-covariance
matrix of the random effects.

For a closer approximations with 200 draws set \code{end = 2100} with
\code{thin = 10} (\emph{better}) or \code{end = 4100} with \code{thin = 20} (\emph{best}),
at the expense of a linear increase in calculation time. Drawing fewer
than 50 observations is not recommended, and such results are best treated
as \emph{indicative}.

It is possible to draw multiple imputations for a subset of parameter draws.
The \code{thin_imp} should be specified as a multiple of \code{thin}. For example,
if \code{thin_imp} is 10 times \code{thin}, then the procedure will calculate and
store multiple imputations in every tenth of the parameter draws. Thus,
for 200 parameter draws there will be 20 draws from the posterior
distribution of the outcome variable.

The total number of parameter draws equals \code{(end - start + 1) / thin}. The number
of multiple imputations equals \code{(end - start + 1) / thin_imp}. By default,
\code{thin_imp} is \code{Inf} and does not produce imputations.
}
\references{
Kasim RM, Raudenbush SW. (1998). Application of Gibbs sampling to nested
variance components models with heterogeneous within-group variance. Journal
of Educational and Behavioral Statistics, 23(2), 93--116.
}
\author{
Stef van Buuren, based on \code{\link[mice:mice.impute.2l.norm]{mice::mice.impute.2l.norm()}}
}
