% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.basis.R
\name{make_basis}
\alias{make_basis}
\title{Create linear splines basis}
\usage{
make_basis(x, knots = round(c(0, 28/365.25, 56/365.25, 1/4, 1/3, 1/2, 7.5/12,
  9/12, 11/12, 14/12, 18/12, 2), 4), boundary = c(0, 3), degree = 1)
}
\arguments{
\item{x}{a vector of ages at which the basis function are needed}

\item{knots}{a vector of internal knots}

\item{boundary}{vector of external knots}

\item{degree}{the degree of the spline. The broken stick model
requires linear splines, so the default is \code{degree = 1}.}
}
\value{
A matrix with \code{length(x)} rows and \code{length(breaks)}
columns, with some extra attributes described by \code{bs()}.
}
\description{
This function creates the basis function of a linear first-order splines
at a user-specific set of break points. The default knots 
code the age range 0-3 years.
}
\author{
Stef van Buuren, 2016
}

