% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula for brokenstick model}
\usage{
parse_formula(f)
}
\arguments{
\item{f}{formula object}
}
\value{
A \code{list} with elements \code{y}, \code{x} and \code{z}.
Each element has length 1.
}
\description{
A bare bones formula parser to extract variables names from
formulas of \code{y ~ x | z}. It return the name of
the first variable mentioned in each formula component.
}
\examples{
# examples that yield identical result
parse_formula(y ~ x | z)
parse_formula(y + a ~ x + x1 | z + b)
parse_formula(y + a + log(b) ~ x + x1 * c | z + d)
}
