% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brokenstick}
\alias{plot.brokenstick}
\title{Plot observed and fitted trajectories}
\usage{
\method{plot}{brokenstick}(x, py, px, ids = NULL, max_ids = 3,
  x_trim = c(-Inf, Inf), measurements = TRUE, estimates = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{brokenstick} or \code{brokenstick_export}}

\item{py}{A vector with measurements using the same response scale as the
fitted model. Used as \code{y} in \code{\link{predict.brokenstick}}.}

\item{px}{A vector with decimal ages of length \code{length(py)}.
fitted model. Used as \code{x} in \code{\link{predict.brokenstick}}.}

\item{ids}{A vector with one or more subject identification codes.}

\item{max_ids}{A scalar indicating the number of individual plots. The default is 3, which plots the trajectories of the first three persons. The \code{max_ids} is a safety measure to prevent unintended plots of the entire data set.}

\item{x_trim}{A range on the x-axis that can be used to subset values that are
displayed}

\item{measurements}{A logical indicating whether the measurements should be plotted. The default is \code{TRUE}.}

\item{estimates}{A logical indicating whether the broken stick estimates should be plotted.
By default, it is \code{TRUE} if there are internal knots, and \code{FALSE} if there are
only boundary knots.}

\item{...}{Extra arguments passed down to \code{\link[rbokeh]{figure}} and
\code{\link[rbokeh]{ly_lines}}, \code{\link[rbokeh]{ly_points}},
'\code{\link[hbgd]{ly_zband}} and '\code{\link[rbokeh]{grid_plot}} functions.}
}
\value{
An object of class \code{rbokeh}.
}
\description{
The \code{plot} method for a \code{brokenstick} object plots the observed and
fitted trajectories of one or more individuals.
}
\examples{
library("brokenstick")
dat <- smocc_hgtwgt
# fit one line model for data exploration
fit <- brokenstick(y = dat$htcm, x = dat$age, subj = dat$subjid)

# plot measurements for first three cases
plot(fit, zband = FALSE, est = FALSE)
plot(fit, zband = FALSE, est = FALSE, pkg = "bokeh", width = 250, height = 350)

# fit model with knots at 1 and 2 years
fit <- brokenstick(y = dat$haz, x = dat$age, subj = dat$subjid, knots = 1:2)

# ggplot
plot(fit, xlim = c(0, 2.2), ylim = c(-3, 3))
plot(fit, ids = c(10005, 10012), xlim = c(0, 2.2))

# bokeh plots with using x_trim
plot(fit, pkg = "bokeh", width = 250, height = 350, x_trim = c(0, 2.2))

# bokeh plots with xlim and ylim
plot(fit, xlim = c(0, 2), ylim = c(-3, 3), pkg = "b", width = 250, height = 350)
plot(fit, ids = 10005, pkg = "bokeh", width = 350, height = 300)
}
\author{
Stef van Buuren 2017
}
