% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brokenstick}
\alias{plot.brokenstick}
\title{Plot observed and fitted trajectories}
\usage{
\method{plot}{brokenstick}(x, py, px, ids = NULL, max_ids = 3,
  x_trim = c(-Inf, Inf), show_measurements = TRUE, show_estimates = TRUE,
  ...)
}
\arguments{
\item{x}{An object of class \code{brokenstick} or \code{brokenstick_export}}

\item{py}{A vector with measurements using the same response scale as the
fitted model. Used as \code{y} in \code{\link{predict.brokenstick}}.}

\item{px}{A vector with decimal ages of length \code{length(py)}.
fitted model. Used as \code{x} in \code{\link{predict.brokenstick}}.}

\item{ids}{A vector with one or more subject identification codes.}

\item{max_ids}{A scalar indicating the maximum number of individual plots. The default is 3, which select the first six figures. Setting \code{max_ids} to a high number increses the maximum. The parameter implements a safety measure to avoid unintended plots of the entire data set.}

\item{x_trim}{A range to select the x-axis}

\item{show_measurements}{A logical indicating whether the measurement should be plotted. The default is \code{TRUE}.}

\item{show_estimates}{A logical indicating whether the estimates should be plotted. The default is \code{TRUE}.}

\item{...}{Extra arguments passed down to \code{\link[rbokeh]{figure}} and
\code{\link[rbokeh]{ly_lines}}, \code{\link[rbokeh]{ly_points}},
'\code{\link[hbgd]{ly_zband}} and '\code{\link[rbokeh]{grid_plot}} functions.}
}
\value{
An object of class \code{rbokeh}.
}
\description{
The \code{plot} method for a \code{brokenstick} object plots the observed and
fitted trajectories of one or more individuals.
}
\examples{
data <- smocc_hgtwgt
head(data)

# fit one line model on data range of length (in cm)
fit <- brokenstick(y = data$htcm, x = data$age, subj = data$subjid)

# plot data and estimates for person 10001
plot(fit, ids = 10001)

# fit one line model on length (in SDS)
fit <- brokenstick(y = data$haz, x = data$age, subj = data$subjid)
plot(fit, ids = 10001, show_zband = TRUE)

smc <- smocc_hgtwgt
fit <- brokenstick(y=smc$haz, x=smc$age, subj = smc$subjid,
 knots = 0:2, boundary = c(0, 3))
plot(fit, ids = c(10001, 10004, 10005),
 px = get_knots(fit), x_trim = c(0, 2.2), xlim = c(0, 2.2))
}
\author{
Stef van Buuren 2016
}
