% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot observed and fitted trajectories from fitted brokenstick model}
\usage{
plot_trajectory(x = x, data = data, color.y = c("blue", "darkgreen"),
  size.y = 10, color.yhat = c("red", "darkgreen"), size.yhat = 10,
  height = 520, width = 480, nrow = 1, x_range = c(0, 2), ...)
}
\arguments{
\item{x}{An object of class \code{brokenstick} or \code{brokenstick_export}}

\item{data}{A list of data frames produced by \code{predict.brokenstick} and \code{predict.brokenstick_export}}

\item{color.y}{A character vector with two elements specifying the symbol and line color of the measured data points}

\item{size.y}{Dot size of measured data points}

\item{color.yhat}{A character vector with two elements specifying the symbol and line color of the predicted data points}

\item{size.yhat}{Dot size of predicted data points}

\item{height}{Figure height in pixels}

\item{width}{Figure width in pixels}

\item{nrow}{Number of rows in plot}

\item{x_range}{a vector specifying the range (min, max) that the superposed growth standard should span on the x-axis}

\item{\dots}{Parameters passed down to \code{\link[rbokeh]{figure}}, 
\code{\link[rbokeh]{ly_lines}}, \code{\link[rbokeh]{ly_points}} 
and '\code{\link[rbokeh]{grid_plot}} functions.}
}
\value{
An object of class \code{rbokeh}
}
\description{
This function is called by \code{plot.brokenstick()} and \code{plot.brokenstick_export()}, and implements the plotting routines using \code{rbokeh}.
}

