% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot observed and fitted trajectories from fitted brokenstick model}
\usage{
plot_trajectory(
  x,
  data,
  color_y = c("blue", "grey"),
  size_y = 2,
  color_yhat = c("red", "grey"),
  size_yhat = 2,
  ncol = 3L,
  xlab = "Age (years)",
  ylab = "Length (SDS)",
  zband = TRUE,
  zband_range = NULL,
  xlim = NULL,
  ylim = NULL,
  theme = ggplot2::theme_light()
)
}
\arguments{
\item{x}{An object of class \code{brokenstick}.}

\item{data}{A \code{data.frame} with columns names \code{x$names}, \code{".source"}
and \code{".pred"}.}

\item{color_y}{A character vector with two elements specifying the symbol and line color of the measured data points}

\item{size_y}{Dot size of measured data points}

\item{color_yhat}{A character vector with two elements specifying the symbol and line color of the predicted data points}

\item{size_yhat}{Dot size of predicted data points}

\item{ncol}{Number of columns in plot}

\item{xlab}{The label of the x-axis}

\item{ylab}{The label of the y-axis}

\item{zband}{A logical indicating whether the Z-score band should be
added to the plot. The default is \code{TRUE}.}

\item{zband_range}{a vector specifying the range (min, max) that the superposed growth standard should span on the x-axis. The
default is the entire data range.}

\item{xlim}{Vector of length 2 with range of x-axis}

\item{ylim}{Vector of length 2 with range of y-axis}

\item{theme}{Plotting theme}
}
\value{
An object of class \code{ggplot}
}
\description{
This function plot the observed and fitted trajectories from brokenstick model.
Actual plotting may be done by \code{gglot}.
}
\examples{
# plot first three cases
# plot(fit_200, brokenstick::smocc_200)
}
\seealso{
\link{plot.brokenstick}
}
