% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.brokenstick.export}
\alias{predict.brokenstick.export}
\title{Predict growth curve according to the broken stick model}
\usage{
\method{predict}{brokenstick.export}(object, y, age, ...)
}
\arguments{
\item{object}{An object of class \code{brokenstick.export}
containing the estimated parameters of the broken stick model}

\item{y}{A vector with measurements}

\item{age}{A vector with decimal ages of length \code{length(y)}}

\item{\dots}{Additional arguments (not used)}
}
\value{
A data frame with \code{length(slot(object, "knots"))
+ slot(object, "degree")} elements with 
predicted values
}
\description{
For a given child, extract the conditional modes of the
random effects as specified by the broken stick (=linear
piecewise spline) model.
}
\author{
Stef van Buuren, 2015
}

