% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokenstick-predict.R
\name{predict_new}
\alias{predict_new}
\alias{predict_new.brokenstick}
\title{Predict from a \code{brokenstick}}
\usage{
predict_new(object, new_data, type = "numeric", ...)

\method{predict_new}{brokenstick}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{brokenstick}
}
\examples{
train <- mtcars[1:20,]
test <- mtcars[21:32, -1]

# Fit
mod <- brokenstick(mpg ~ disp | cyl, train)

# Predict, with preprocessing
predict_new(mod, test)

}
