% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.brokenstick}
\alias{residuals.brokenstick}
\title{Extract residuals from brokenstick model}
\usage{
\method{residuals}{brokenstick}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{newdata}{Optional. A data frame in which to look for variables with
which to predict. If omitted and if \code{isFALSE(object$light)}, the fitted
predictors are used.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
See \code{\link[=predict.brokenstick]{predict.brokenstick()}}.
}
\description{
Extract residuals from brokenstick model
}
\seealso{
Other brokenstick: 
\code{\link{fitted.brokenstick}()}
}
\concept{brokenstick}
