% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.brokenstick}
\alias{residuals.brokenstick}
\title{Extract residuals from brokenstick model}
\usage{
\method{residuals}{brokenstick}(object, new_data = NULL, ...)
}
\arguments{
\item{object}{A \code{brokenstick} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{\dots}{Additional arguments. Ignored.}
}
\value{
See \code{\link[=predict_new.brokenstick]{predict_new.brokenstick()}}.
}
\description{
Extract residuals from brokenstick model
}
\seealso{
Other brokenstick: 
\code{\link{fitted.brokenstick}()}
}
\concept{brokenstick}
